Se agregan configuracines docker para mysql
