+++ MainManagementApp/.idea/encodings.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/CommonLibrary" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/Management" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/ProductApp" charset="UTF-8" />
   </component>
