+++ MainManagementApp/CommonLibrary/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+  <installed facet="cloudfoundry.standalone.app" version="1.0"/>
+</faceted-project>
