+++ MainManagementApp/CommonLibrary/src/main/java/com/stack/jspdemo/App.java
@@ -1,0 +1,13 @@
+package com.stack.jspdemo;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
