+++ MainManagementApp/Management/pom.xml
@@ -4,7 +4,6 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-
 	<artifactId>Management</artifactId>
 
 	<packaging>jar</packaging>
@@ -17,8 +16,13 @@
 		<artifactId>MainManagementApp</artifactId>
 		<version>0.0.1-SNAPSHOT</version>
 	</parent>
-
 	<dependencies>
+		<dependency>
+			<groupId>com.stack</groupId>
+			<artifactId>CommonLibrary</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
+
 		<dependency>
 			<groupId>javax.servlet</groupId>
 			<artifactId>jstl</artifactId>
