+++ MainManagementApp/Management/src/main/java/com/stack/jspdemo/rabbit/consumer/Consumer.java
@@ -1,19 +1,6 @@
 package com.stack.jspdemo.rabbit.consumer;
 
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
-
-import org.springframework.amqp.core.Message;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.stack.jspdemo.model.Product;
-import com.stack.jspdemo.service.ProductService;
 
 @Component
 public class Consumer {
