+++ MainManagementApp/Management/src/main/java/com/stack/jspdemo/rabbit/producer/Producer.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.stack.jspdemo.model.Product;
+import com.stack.jspdemo.model.ProductDTO;
 
 @Component
 public class Producer {
@@ -19,7 +19,7 @@ public class Producer {
 	@Value("${jsa.rabbitmq.routingkey}")
 	private String routingkey;
 
-	public void produce(Product product) {
+	public void produce(ProductDTO product) {
 		
 		amqpTemplate.convertAndSend(exchange, routingkey, product);
 		System.out.println("Send msg = " + product);
