+++ MainManagementApp/Management/src/main/java/com/stack/jspdemo/service/ProductService.java
@@ -6,7 +6,6 @@ import java.util.List;
 
 import org.springframework.amqp.core.Message;
 import org.springframework.amqp.core.MessageListener;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,7 +13,7 @@ import org.springframework.stereotype.Service;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.stack.jspdemo.model.Product;
+import com.stack.jspdemo.model.ProductDTO;
 import com.stack.jspdemo.rabbit.consumer.Consumer;
 import com.stack.jspdemo.rabbit.producer.Producer;
 
@@ -34,15 +33,15 @@ public class ProductService implements MessageListener {
 	@Autowired
 	RabbitTemplate rabbitTemplate;
 
-	public static List<Product> products = new ArrayList<>();
+	public static List<ProductDTO> products = new ArrayList<>();
 
-	public Product addProduct(Product product) {
+	public ProductDTO addProduct(ProductDTO product) {
 		producer.produce(product);
 		products.add(product);
 		return product;
 	}
 
-	public List<Product> findAll() {
+	public List<ProductDTO> findAll() {
 		System.out.println(products.toString());
 		return products;
 	}
@@ -54,8 +53,8 @@ public class ProductService implements MessageListener {
 
 		byte[] body = message.getBody();
 		String s = new String(body);
-		Product[] p = mapper.readValue(s, Product[].class);
-		for (Product product : p) {
+		ProductDTO[] p = mapper.readValue(s, ProductDTO[].class);
+		for (ProductDTO product : p) {
 			products.add(product);
 		}
 		// message.getMessageProperties().getReplyTo();
@@ -70,10 +69,10 @@ public class ProductService implements MessageListener {
 
 		byte[] body = message.getBody();
 		String s = new String(body);
-		Product[] p;
+		ProductDTO[] p;
 		try {
-			p = mapper.readValue(s, Product[].class);
-			for (Product product : p) {
+			p = mapper.readValue(s, ProductDTO[].class);
+			for (ProductDTO product : p) {
 				// System.out.println(p);
 				products.add(product);
 			}
@@ -85,8 +84,8 @@ public class ProductService implements MessageListener {
 	}
 
 	//@RabbitListener(queues = "${jsa.rabbitmq.queue2}", containerFactory = "jsaFactory2")
-	public List<Product> printAllProducts() {
-		for (Product product : products) {
+	public List<ProductDTO> printAllProducts() {
+		for (ProductDTO product : products) {
 			System.out.println("Desde metodo printAllProducts" + product);
 		}
 		return products;
