+++ MainManagementApp/ProductApp/docker-compose.yml
@@ -1,0 +1,35 @@
+version: '2.3'
+services:
+  springappserver:
+    build:
+      context: .
+      dockerfile: springapp.dockerfile
+    ports:
+    - 8080:8080
+    networks:
+    - net-spring-db
+    volumes:
+    - .:/vol/development
+    depends_on:
+    - mysqldbserver
+  mysqldbserver:
+    build:
+      context: .
+      dockerfile: mysqldb.dockerfile
+    ports:
+    - 3306:3306
+    networks:
+    - net-spring-db
+    environment:
+      MYSQL_DATABASE: ManagementProduct
+      MYSQL_USER: myuser
+      MYSQL_PASSWORD: mypassword
+      MYSQL_ROOT_PASSWORD: myrootpassword
+    container_name: mysqldbserver
+    volumes:
+    - type: bind
+      source: ./mysql
+      target: /docker-entrypoint-initdb.d
+networks:
+  net-spring-db:
+    driver: bridge
