+++ MainManagementApp/ProductApp/src/main/java/com/stack/jspdemo/rabbit/producer/Producer.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.stack.jspdemo.model.Product;
+import com.stack.jspdemo.entity.Product;
 
 @Component
 public class Producer {
