+++ MainManagementApp/ProductApp/src/main/java/com/stack/jspdemo/repository/ProductRepository.java
@@ -2,7 +2,7 @@ package com.stack.jspdemo.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.stack.jspdemo.model.Product;
+import com.stack.jspdemo.entity.Product;
 
 public interface ProductRepository extends CrudRepository<Product, Long>, ProductRepositoryCustom{
