+++ MainManagementApp/ProductApp/src/main/java/com/stack/jspdemo/repository/ProductRepositoryCustom.java
@@ -2,7 +2,7 @@ package com.stack.jspdemo.repository;
 
 import java.util.List;
 
-import com.stack.jspdemo.model.Product;
+import com.stack.jspdemo.entity.Product;
 
 public interface ProductRepositoryCustom {
 	List<Product> getAllProducts();
