+++ MainManagementApp/ProductApp/src/main/resources/application.properties
@@ -2,8 +2,8 @@ spring.application.name=ProductApp
 server.port=8000
 
 spring.datasource.url=jdbc:mysql://localhost:3306/ManagementProduct
-spring.datasource.username=jarenas
-spring.datasource.password=jesus1985
+spring.datasource.username=myuser
+spring.datasource.password=mypassword
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
 
 spring.jpa.show-sql=true
@@ -15,13 +15,9 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialec
 logging.level.org.org.hibernate.SQL=debug
 logging.level.org.org.springframework.web=DEBUG
 
-#spring.rabbitmq.host=localhost
-#spring.rabbitmq.port=5672
-#spring.rabbitmq.username=guest
-#spring.rabbitmq.password=guest
 #Consumer a product
 jsa.rabbitmq.queue=jsa.queue
-#logging.file=jsa-app.log
+logging.file=jsa-app.log
 
 jsa.rabbitmq.exchange3=TestExchange
 jsa.rabbitmq.routingkey3=TestRouting
